[Add]: Implement create and edit task #issu7&8

+++ src/main/java/com/example/taskmanager/service/task/TaskServiceImpl.java
@@ -1,25 +1,49 @@
 package com.example.taskmanager.service.task;
 
 import com.example.taskmanager.repository.task.TaskRepository;
+import com.example.taskmanager.repository.label.LabelRepository;
+import com.example.taskmanager.repository.category.CategoryRepository;
 import org.springframework.stereotype.Service;
 import com.example.taskmanager.model.Task;
 import com.example.taskmanager.model.User;
+import com.example.taskmanager.model.Label;
+import com.example.taskmanager.model.Category;
 import com.example.taskmanager.model.CustomUserDetails;
-import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import com.example.taskmanager.requests.TaskFormRequest;
+import org.springframework.web.multipart.MultipartFile;
+import jakarta.persistence.EntityNotFoundException;
+import com.example.taskmanager.utils.FileUtils;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.Optional;
 
 @Service
 public class TaskServiceImpl implements TaskService {
 
-    private final TaskRepository taskRepository;
+    private final String UPLOAD_DIR = "uploads/";
+    private final Integer IS_DELETE_FILE = 1;
 
-    public TaskServiceImpl(TaskRepository taskRepository) {
+    private final TaskRepository taskRepository;
+    private final LabelRepository labelRepository;
+    private final CategoryRepository categoryRepository;
+
+    public TaskServiceImpl(
+        TaskRepository taskRepository,
+        CategoryRepository categoryRepository,
+        LabelRepository labelRepository
+    ) {
         this.taskRepository = taskRepository;
+        this.categoryRepository = categoryRepository;
+        this.labelRepository= labelRepository;
     }
 
     public Page<Task> listTask(int pageNumber, String searchValue, String sortBy, String sortValue) {
@@ -47,10 +71,110 @@ public class TaskServiceImpl implements TaskService {
 
     public Boolean delete(Long id) {
         try {
+            Task task = this.detail(id);
+            if(task.getTitle() == null) {
+                return false;
+            }
+
+            if (task.getPhoto() != null) {
+                FileUtils.deleteFileIfExistsNio(task.getPhoto());
+            }
+
             taskRepository.deleteById(id);
             return true;
         } catch (Exception e) {
             return false;
         }
     }
+
+    public Boolean store(TaskFormRequest taskFormRequest) {
+        try {
+            Category categories = categoryRepository.findById(taskFormRequest.getCategoryId()).get();
+
+            List<Long> labelIds = Arrays.asList(taskFormRequest.getLabelId());
+            Set<Label> labels = new HashSet<>(labelRepository.findAllById(labelIds));
+
+            Task task = new Task();
+
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+            CustomUserDetails currentUserDetails = (CustomUserDetails) auth.getPrincipal();
+            User currentUser = currentUserDetails.getUser();
+            task.setUser(currentUser);
+
+            task.setTitle(taskFormRequest.getTitle());
+            task.setDescription(taskFormRequest.getDescription());
+            task.setDueDate(taskFormRequest.getDueDate());
+
+            MultipartFile photo = taskFormRequest.getPhoto();
+            if(!photo.isEmpty()) {
+                task.setPhoto(FileUtils.uploadFile(photo, UPLOAD_DIR));
+            }            
+
+            task.setStatus(taskFormRequest.getStatus());
+            task.setPriority(taskFormRequest.getPriority());
+            task.setCategory(categories);
+            task.setLabels(labels);
+
+            taskRepository.save(task);
+
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    public Boolean update(Long id, TaskFormRequest taskFormRequest) {
+
+        Task task = this.detail(id);
+        if(task.getTitle() == null) {
+            return false;
+        }
+
+        try {
+            if(
+                task.getPhoto() != null 
+                && taskFormRequest.getDeleteFile() == IS_DELETE_FILE
+            ) {
+                FileUtils.deleteFileIfExistsNio(task.getPhoto());
+            }
+
+            Category categories = categoryRepository.findById(taskFormRequest.getCategoryId()).get();
+
+            List<Long> labelIds = Arrays.asList(taskFormRequest.getLabelId());
+            Set<Label> labels = new HashSet<>(labelRepository.findAllById(labelIds));
+
+            task.setTitle(taskFormRequest.getTitle());
+            task.setDescription(taskFormRequest.getDescription());
+            task.setDueDate(taskFormRequest.getDueDate());
+
+            MultipartFile photo = taskFormRequest.getPhoto();
+            if(!photo.isEmpty()) {
+                task.setPhoto(FileUtils.uploadFile(photo, UPLOAD_DIR));
+            } 
+            
+            if(photo.isEmpty() && taskFormRequest.getDeleteFile() == IS_DELETE_FILE) {
+                task.setPhoto(null);
+            }
+
+            task.setStatus(taskFormRequest.getStatus());
+            task.setPriority(taskFormRequest.getPriority());
+            task.setCategory(categories);
+            task.setLabels(labels);
+
+            taskRepository.save(task);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    public Task detail(Long id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        CustomUserDetails currentUserDetails = (CustomUserDetails) auth.getPrincipal();
+        User currentUser = currentUserDetails.getUser();
+
+        Optional<Task> task =  taskRepository.findFirstByUserAndId(currentUser, id);
+
+        return task.orElseThrow(() -> new EntityNotFoundException("Task with ID " + id + " not found"));
+    };
 }
